+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - `compare` method to score predictions relative to each other
 
+### Fixed
+- Update [plot script](plot.py) to work with new API
+
 ## [0.2.0] - 2020-05-15
 
 ### Added